+++ Sources/LiveViewNative/Property Wrappers/FormState.swift
@@ -66,6 +66,14 @@ public struct FormState<Value: FormValue> {
     @Environment(\.formModel) private var formModel: FormModel?
     @Environment(\.coordinatorEnvironment) private var coordinator
     
+    /// Use this with the ``SwiftUI/View/focused(_:)`` modifier to prevent server-side updates while editing.
+    @FocusState public var isFocused: Bool
+    
+    /// Use this with `onEditingChanged` callbacks to prevent server-side updates while editing.
+    ///
+    /// Prefer ``isFocused`` for elements that can be focused.
+    @State public var isEditing: Bool = false
+    
     @Event("phx-change", type: "form") private var changeEvent
     
     /// Creates a `FormState` property wrapper with a default value that will be used when no other value is present.
@@ -174,10 +182,19 @@ public struct FormState<Value: FormValue> {
     }
     
     private func resolveMode() {
+        let elementName = element.attributeValue(for: "name")
         if case .unknown = data.mode {
             if let formModel {
-                if let elementName = element.attributeValue(for: "name") {
-                    data.setFormModel(formModel, elementName: elementName)
+                if let elementName {
+                    data.bind(
+                        formModel,
+                        to: _element,
+                        elementName: elementName,
+                        attribute: valueAttribute,
+                        defaultValue: defaultValue,
+                        isFocused: $isFocused,
+                        isEditing: $isEditing
+                    )
                     formModel.setInitialValue(initialValue, forName: elementName)
                     data.mode = .form(formModel)
                 } else {
@@ -222,11 +239,43 @@ extension FormState: DynamicProperty {
 private class FormStateData<Value: FormValue>: ObservableObject {
     var mode: Mode = .unknown
     private var cancellable: AnyCancellable?
+    private var elementCancellable: AnyCancellable?
+    private var focusCancellable: AnyCancellable?
     
-    func setFormModel(_ formModel: FormModel, elementName: String) {
+    func bind(
+        _ formModel: FormModel,
+        to element: ObservedElement,
+        elementName: String,
+        attribute: AttributeName,
+        defaultValue: Value,
+        isFocused: FocusState<Bool>.Binding,
+        isEditing: Binding<Bool>
+    ) {
+        // Trigger a View update when the field's value changes.
         cancellable = formModel.formFieldWillChange
             .filter { $0 == elementName }
             .sink { [unowned self] _ in self.objectWillChange.send() }
+        
+        // When the element updates from the server, sync the new value into the form.
+        elementCancellable = element.projectedValue
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self, weak formModel] _ in
+                // ignore server updates if the field is focused.
+                guard !isFocused.wrappedValue && !isEditing.wrappedValue else { return }
+                formModel?.setServerValue(
+                    element.wrappedValue.attribute(named: attribute)
+                        .flatMap { Value.fromAttribute($0, on: element.wrappedValue) }
+                        ?? defaultValue,
+                    forName: elementName
+                )
+            }
+        
+        // Remove all focus from form fields when the form is submitted.
+        focusCancellable = formModel.formWillSubmit
+            .sink { _ in
+                isFocused.wrappedValue = false
+                isEditing.wrappedValue = false
+            }
     }
     
     enum Mode {